Always give the event target as the first listener parameter
To be consistent, always give the object that dispatches an event as the
first listener parameter. In some cases this will break code as existing
even paramters are moved by one.

Also given the options object in animation events directly as the second
parameter, to be consistend with change events.

Change-Id: Ic92d481df9bc982571311ee8e9fade98ee6d5156