Fix broken selection events for Button-like and Action-like widgets
The events didn't work because I copy-pasted code from CollectionView
where the selection event is written with a lower-case "s" to
Button/Action, where it is written with upper-case "S". 

There where no tests for the listen ops before, so tests didn't go red. 

iOS is case-insensitive for the first character, so the exmples worked
when tested on this platform.

In Android the Actions still worked because the listen operation is
ignored for these Widget, CollectionView is lower case anyway, and
buttons were presumably not tested on this platform.

Add listen tests and fix the lower-case "s".

Change-Id: Ib2fdde56e5d17f5d554fc632af4bc047ff67dd21