Revise and align SearchAction and TextInput events
SearchAction `submit` is now `accept`. Second parameter is the query.
SearchAction `modify` is now `input`. Second parameter is the query.

TextInput `accept` now has a second parameter, which is the text.
TextInput `input` has been added as an alternative to `change:text`. It
has the same parameter, but fires only on user changes.

The old events are still available.

Change-Id: Ic80a6a01e884638fd6b270cf48482b19b3066130