Fix disposal of child widgets
It turned out that after disposing a widget, its children had been
released from the proxy store but their isDisposed() method still
returned false. That was because the dispose method was not used
for children (which do not require the native destroy operation).

Create an internal _dispose method that can be used with a boolean
parameter to skip the native destroy operation. Use this method for
disposing both, parent and children.

Instead of _destroy, introduce a _release method that subtypes can
overwrite to release resources on disposal. Widget uses _releaase
to unhook from the parent and dispose of the children.

Move tests for parent/child disposal to Widget, as Proxy does not
have concept of parent-child relationship.

Change-Id: Iadfc58bbed372d89788ebde293967352dc47bbb0