Remove support for ShowPage event
The ShowPage used to be triggered when a page has been selected in the
client-side page switcher, which has been replaced by the Drawer and
PageSelector. The only case for page transistions triggered on the
client is now back-navigation which is covered by the ShowPreviousPage
event.

iOS also sends this event when a page transition is finished, however,
since animated page transition are an implementation detail, they are
not relevant for the JS side.

Change-Id: Ia0d193febb20c602b891ee7d2452aedfa9dbfb6d