Close stacked pages when opening / closing new pages
The JS code is responsible for the management of pages. This includes
cleaning up pages that are no longer needed. There are different cases
that must be considered, according to the rules laid out in doc/ui.md.

Dispose calls for old pages must be sent *after* setting the new
activePage in order to avoid client-side problems.

Fix #263

Change-Id: If8a991f9958b53410f8315daaeb9f1b1d3f5bf5d