Remove "known issues" segments from the documentation
The described issues are largely resolved, or will be treated as normal
bugs. 

In case of the transformation issue, it will be treated and documented
as an accepted platform quirk. 

The one-instance limit for iOS web-views is not enforced and is not
known to cause any ciritical problems. It's therefore now documented as
a recommendation.

Fix #385
Fix #386
Fix #387

Change-Id: Iba9ec8418e40c202fe1d3abb9a39a2bd90104ad9