Throw error when opening a page without a top-level page
The pages model demands that stacked pages can only be opened on top
of a top-level page. That's because a stacked page displays a button
for back navigation that can only work if there is a root page.

Opening a stacked page without a top-level page will lead to a broken
navigation. It's better to fail than to continue.

Adjust tests to avoid this error.

See #404

Change-Id: Ia7bfee56a88b80187d1955f7ee39c2042ddba7fb