Remove return value from animate method
The practice of chaining animationend listeners to the animate method
must be considered harmful, as it will add a new listener everytime the
animation is triggered, which is most likely not the intention.

Example:

    button.on("selection", function() {
      view.animate({
        opacity: 0
      }, {
        duration: 1000
      }).on("animationend", function() {
        view.set("text", "done");
      });
    });

If a new animation is triggered in the animationend listener, this can
easily lead to en endless loop. Since we've even found this pattern in
our own examples, it seems like a dangerous pitfal in our API.

To avoid this pattern, don't let the animate method return the widget
anymore. This will also open up the API for the introduction of
promises.

Change-Id: Iacb7a8197cef3d0149f8bc13bbf3ef61c65aa2ab