Extract API reference for non-widgets objects from .md files to json
APIs should all be documented in the same recognizable format.
Previously the widgets documentation was generated from json, with all
other APIs being described in various formats in the .md files.

This change extracts these API descriptions from from the .md files and
adds them to json definitions. During the build they are merged with the
.md again.

Change-Id: Ia85c07d52fee268886ae32116f995d5a9ef966f8