Use caret version ranges in snippets and examples
In node-semver, a dependency to "1.0.0" depends on the *exact* version,
that is, patch versions would not be found. The caret prefix creates
a range that includes all version up to, but not including, the next
major.

According to `npm help install`, the caret is npm's default semver
range operator, it's also the version added to the package.json by
`npm install --save`.

https://github.com/npm/node-semver#caret-ranges-123-025-004

Change-Id: I27f22190fb52c96061d97235436b66e600487665