Introduce installPatch function on app object
Allow to patch apps to apply critical fixes without having to wait
for an update becoming available in the app stores. Example:

    tabris.app.installPatch(patchUrl, function(error, patch) {
      if (error) {
        // show error dialog
      } else {
        // confirm reload
        tabris.app.reload();
      }
    });

The clients will download a zip file from the given patchUrl, extract
it, and prepare to overlay the app's resources with the contents of
the patch zip after the next reload.

After successful installation, the callback will be called with the
parameter `patch` containing the contents of the file `patch.json`
if it was present in the patch. In case of an error, the callback
will be called with the `error` parameter being set.

See #479

Change-Id: I8919f83ed33f7fc48d480c6123dd107b8c05cf32