Ignore patch version in version check, add tests
We should be able to provide newer patch versions on npm to fix
critical issues on the JS side. Those patch versions should not
introduce new features and they should not require a newer version
of the native clients. Hence patch versions should not affect
compatibility between clients and JS.

Add a unit test for bootstrap.js.

Rename module.html to boot.html and include the bootstrap tests.
As a result, the html files to run jasmine tests in the browser
match the jasmine tasks in the grunt build ("boot" and "tabris").

Move both boot tests to a subfolder to safely exclude them from the
tabris test task.

Fix #477

Change-Id: If6a5f4652de3be1420095edce6d02cb0e3ed6026