Support different cell types in CollectionView
Different cell types require different item heights. Since it both
platforms can also support different heights on individual items,
add support for this as well. However, iOS demands that item heights
are determined before items are created or recycled.

Extends the existing property `itemHeight` to also support functions.
If set to a function, don't pass the property directly to the client.
Call the function for every item in the `requestinfo` event instead
and pass the results in the `descibeItem` call.

Introduce a new optional property `cellType` that can be set to a
function. This function returns the cell type for a given item.

#131

Change-Id: I800f5d83c8600f4d42f10782e5db0cade8c81afe