Add support for class selectors
Class selectors now work as in CSS, except that currently only one class
can be selected, no combinations of multiple classes or class and type
or class and id. 

The classes are stored in an array directly attached to the widget
called "classList", inspired by the DOM property of the same name. This
should eventually become public API, but not before we have adjusted it
to be a subset of the W3C API, i.e. not an actual array but an
array-like object.

Change-Id: I4a29025a0e534a720d7546f8a3922ae5bcd0ee97