Add support for the Switch Widget
The switch widget is a button with two states, such as
on or off. States can be changed by sliding the switch from
side to side.

Since the functionality is very similar to ToggleButton and CheckBox,
provide the same API, i.e. a property `selection` and an event named
`select`.

For the native bridge protocol, use the event name `toggle` instead
of `select` which seems more future-oriented (see #515).

Change-Id: I3a988970d72711352b73545a256956d3b6d77278