Prevent object is disposed error in once()
When a widget is disposed inside an event listener, other listeners
attached to the same event type will still be notified of this event.
This behavior is consistent with removing event listener from within a
listener.

In case of once(), the framework code calls off() internally. When the
widget is disposed by another listener on the same event, it can happen
that the widget is already disposed and calling off() will result in an
"Object is disposed" error.

Prevent this error by omitting the listener removal if disposed.
In this case, all listeners have already been removed.

Fix #531

Change-Id: Icd993b81d7172617a48836dcdc72a42ef2dadf37