Fix selector support for TabFolder
When searching the TabFolder with the `children` or `find` method, the
selector was previously either not applied, or the result also contained
the internal item widgets, or the result was missing tabs and items.
This last issue was actually a bug in ProxyCollection, not TabFolder.

Also, the TabFolder `children` method returned an array, not a
ProxyCollection. 

Fix #367

Change-Id: I75e91e09a5b8e94bb8b86834a9427c6599db2672