Put notes, hints and remarks in blockquotes with exclamation icon
Smaller sections of the documentation that deal with important
clarifications or general hints usually start with "Note:" or "Please
note that". Omit this phrase and instead put the text in a blockquote to
visually separate it from the rest of the article. Add a exclamation
icon to identify it as a remark, not a quote.

note.png is public domain and taken from 
https://commons.wikimedia.org/wiki/File:Symbol_point_of_order.svg

Change-Id: Id62442fd79644c6b1249799eacfa98f250edc3af