Rework encoding of layoutData
In the encoding step, normalize all layoutData attributes, to store
them in the same format they are sent to the client.
- transform percentage strings to arrays
- complete checks for attribute values

Introduce a decoding step that simplifies certain cases, i.e. replace
- [0, n] -> n
- [n, 0] -> "n%" when n is a number, otherwise
- [x, 0] -> x

Ensure deep defensive copies are created in getter and setter.

Introduce resolveReferences as a replacement for encodeLayoutData.
This function now only translates widgets and widget references to
their cid.

Change-Id: I9bb7ad0b1554769bf3af5e31693bfd77d5e78348