Rename "util" to "_"
The object `util` is used internally in tabris.js to provide a number
of mostly lodash/underscore compatible functions. Rename the object
to `_` to make it clearer that we're using "standard" functions here.

In the future we may decide to include a subset of lodash or underscore
instead of writing these functions ourselves.

Change-Id: I208a829cdc721dd11343288a340afb641bfa26fc