Dynamically resolve references in layoutData
When layoutData contain symbolic references to widgets, a reference
may become valid or invalid or even be resolved to a different widget
when the list of siblings changes. Therefore all references must be
resolved again whenever the set of child widgets in the parent changes.

Introduce a layout queue that holds all parents whose list of children
has changed. Just before flushing nativeBridge operations, re-render
the layoutData of all children of the parents in this list.

Remove the fallback logic and the `force` parameter in
resolveReferences.

As a further optimization, we might track the cids of the resolved
references and check whether they have really changed.
Another optimization would be to distinguish layoutData that contain
references and those that do not. Only those with symbolic references
would need to be re-rendered.

Change-Id: I3fce7cf5bbf53c59bcd9d88a804b868141a7f882