Move custom setter/getter handling from Proxy.js to Properties.js
Instead of overwriting methods from Properties.js (_applyProperty and
_readProperty), the Proxy now registers custom setter/getter for *all*
native properties. 

Also, the custom setter/getter are now called *instead* of accessing the
local property store, not in *addition*. That way the caching logic is
kept out of Properties.js:

If no custom setter/getter are given to the registerType method,  it
adds default setter/getter that handle caching.

If custom setter/getter are given to the registerType method, they have
to take of caching themselves.

The "noache" flag is still supported (by the default setter/getter), but
will be replaced later by the ability to choose from a set of
pre-defined setter/getter.

Change-Id: If33d01c99d51f59489cd70681a8014a4c6883502