Support 'prev()' reference in layoutData
Allow to align widgets relative to their previous widget without
having to track their id.

Introduce the symbolic reference "prev()" (similar to the function
prev() in jQuery) that is dynamically resolved to the widget that
directly preceeds the target widget. This reference can be used just
like a selector string in layoutData.

When there is no preceeding widget, the reference is replaced with
a zero offset.

Fix #545

Change-Id: I006073429054288aca769020624ae57ecbaf0b88