Support separate layout properties
To allow partial layout changes, support additional widget attributes:

  left, right, top, bottom, width, height, centerX, centerY, baseline

These new properties are mapped to the existing layoutData property.

To still be able to check the consistency of layoutData, render all
layoutData in the layout flush when no more changes are expected. As a
side effect, widgets without parents won't render layoutData. This
includes Page, so set layoutData on it with nativeSet.

Also add missing check for left/width/right and top/height/bottom
contradiction as described in the documentation.

Fix #576

Change-Id: I9c25b023f3e60474d0b00022711593e151b7409f