Combine custom setter/getter in object "access"
Group "set" and "get" functions in a object called "access" in property
definitions. This will later enable to replace these functions with a
single string referencing a pre-defined setter/getter pair.

Use the name "access" because both functions control the read and write
access to the property value at runtime.

Change-Id: Ib5b3949a72fe5b1934ec9464691612ea40a5a507