Remove color and font properties from cache if set to "initial"
Colors and fonts can already be reset by setting them to "initial",
however the getter did then return "initial", not the actual color. This
is because to remove a value from the cache it has to be set to
"undefined", but to reset the property on the native side it has to be
set to "null". Fix this by encoding the value to undefined, but use
custom setter to forward it as "null".

Remove the "textColor" and "background" properties from the UI widget 
since they needlessly overwrites the same properties inherited from 
Widget. That way they use the same custom setter.

Fix #394

Change-Id: I6dc8ed43ac13ec55368ee2147ea576d7ba1e0170