Update and revise layout documentation
Add the new direct layout properties to Widget.json. To be able to give
a useful type, add "edge", "dimension" and "offset" to types.md. Also
mention these new properties in layout.md. 

In layout.md, move the explanations of the various layout attribute
types from the top to the documentation of the different attributes. 

Rationale: While this introduces some redundancy, it was previously very
easy to miss that widgets references can be given as selectors. It also
avoids wording like "Some attributes accept".

Fix #594

Change-Id: Iab28cb99ec0821e99c6c0c1ec79400972494a3e2