Introduce open event on tabris.app
The open event is fired when an app is opened by the call of a custom
url. The entire url that has been called is being passed as a
parameter.

Apps need to register the url scheme that they want to open for.
The registration can be done by a cordova plugin.