Implement getResourceLocation on tabris.app
Provide method getResourceLocation on tabris.app. This method returns
a URL that can be used to access resources that are bundled with the
app. Normally, this will be a file URL, but in case of the dev app
it will be http.

While it should be possible to use the cordova filesystem plug-in to
get absolute URLs, one would have to guess a tabris-specific subfolder
that depends on the build method.

Required for #592

Change-Id: I573a571d40846a3cd8cf510be14f2aadf219dafc