Include type field in all API definition files
In order to simplify the generation of API definitions for Tern and
TypeScript, let every API definition file include a `type` field.
Those types that have a single instance, still contain the additional
`object` field, to require minimal changes to the markdown generator.

For `window.json`, chose `Timer` as type, since this file describes
only the timer API.

I've verified that the generated output is fully identical with the
previous version.

Change-Id: I38656f47b05eb24dbc52c3a3472cb4ea87544725