Support data arrays in Picker items property
Allow to attach arrays of arbitrary data items to the Picker's
`items` property. Extract the display text for an item using a
function that is stored in the property `itemText`.

This allows to connect existing data models more easily. When an
item is selected, the property `selection` contains a reference
to the selected data item. That's more convenient than dealing
with the selectionIndex.

With this change, the `items` property matches the semantics of
corresponding CollectionView property.

If we provide data binding support using a `model` property,
`items` could become an alias of `model`.

Fix #554

Change-Id: Ia94dcf57bd66337777b7ae48b95139c5bf2ff41e