Allow for TabFolders without a tab bar
Add "hidden" as a possible value for the property tabBarLocation.

We currently map this property to style flags. Instead of inventing
another style flag, pass the property directly to the clients.

The property is "static", i.e. it can only be set at creation time.
This is not enforced by the code yet, as it is not for other static
properties. We should rather handle this case in the property system.

Change-Id: Ia26c7bd34a19ad7c24232a1cb1118173ffbc627b