Add support for tabris.js-specific TabFolder protocol
In the future the clients should support a TabFolder that matches the
public API instead of the RWT-style API. Until all clients do so both
versions are in the code and loaded depending on the detected client.

Currently the device properties can not be accessed while the tabris
module is being parsed. Until this changes the TabFolder implementations
are wrapped instead a tabris.load call. To ensure these are executed for
the jasmine tests, run tabris._init before starting them.

Change-Id: I24f86bfe19644d0b31771aa0e90ddb6ef5b64474