Keep array properties by reference, not as copy
Making defensive copies of array properties can be confusing, as the
getter would not return the original list, but only a copy. In case
of an array-like, additional methods would not be available.

Making defensive copies of arrays can also affect performance.
For example, when requesting the number of CollectionView items in a
scroll event, every call to collectionView.get("items").length would
result in an array copy operation.

Currently, this change applies only to the items property of Picker
and CollectionView.

Change-Id: I80d6f35d9fca0a1166702042d2e0463332148302