Introduce properties scrollX and scrollY on ScrollView
These properties allows to get and to set the current scroll position
of a ScrollView.

Since a ScrollView can only be either horizontal or vertical, an
alternative would be a single property `scroll` that controls
the scroll position in the ScrollView's scrolling direction.
However, for geometric calculations, separate properties for x and y
may be more convenient. They also fit better with the x and y offset
in the scroll event.

Fix #662

Change-Id: I8bab81e8c245c6e4bb81cb7c29343ab858a668ff