Make SearchAction a widget, add to tabris.ui
Historically, Action and SearchAction haven't been widgets as they
cannot be part of a layout and hence don't support layout data.
Later, we decided to turn Action into a widget, both in the doc and
in the implementation. Since SearchAction inherits from Action,
it should be treated as a widget as well.

Align SearchAction.js with Action.js.

Fix #658

Change-Id: I95d6f0169f514daf369fc06d91da1bf9fdb247e0