Add navigate event to WebView, make preventDefault a function
Allow to intercept requests in WebView by calling event.preventDefault()
in a listener on "navigation" (see #113).

We already introduced an event *field* preventDefault for the
backnavigation event in tabris.app that can be set to true.

In contrast, DOM events have a preventDefault *method*, that "cancels
the event ... without stopping further propagation of the event" [1].
This method will set a field defaultPrevented.

Moreover, setting the field to `true` allows other listener to reset it
to `false`. Code that does so would depend on the order in which
listeners are called.

In order to avoid confusion, turn preventDefault to a method. This also
prevents resetting the flag after the method has been called once.

For compatibility, continue to support setting it to true.

See #113

[1] http://developer.mozilla.org/en-US/docs/Web/API/Event/preventDefault

Change-Id: Ibf97bfa5b15151ccef800cbc64d1e7b537f6b4c8