Simplify URL validation in XMLHttpRequest
A full validation of all possible URL patterns results in an overly
complex regex that turned out too expensive to process.

Replace commented code with a basic check for supported URL schemes.
Accept URLs without a scheme, as those will be handled as file URLs.
Leave the rest of the validation to the clients.

Change-Id: Ie4aa8362b5223f9efc8dc249fa171a48e77336ff