Enable Backbone section in tests
We have some optional unit tests to ensure API-compatibility of our
Events implementation with Backbone. These tests are only executed
when Backbone is available in the global scope.

Since this is normally not the case, this conditional execution lead
to not running these tests at all anymore.

Add backbone as a dev dependency and make it available to both,
browser-based and automated tests.

Remove condition in Events.spec.js, move one tabris-specific test
case to the tabris-only section.

Change-Id: I3944354d952396896caee9c2c64c5a2b7a906add