Don't use tabris function to create singletons
Some native objects are singletons, those do not require a create
operation, they are referenced by a fixed cid that is known by the
clients.

Distinguish singleton types from normal types by using a `_cid` field
in registerType instead of the `_type` field. This removes some magic
from the tabris() function. It also removes the hidden support for a
string argument to pass the cid to a constructor.

Change-Id: I881a2f4e4e7478270c25f2a5b148949ab976884c