Revise super calls in widgets
We use the util method `extendPrototype` to build up prototype chains.
This method adds a function to walk up the prototype chain and call a
method in the super prototype.

Prefix this method with `_`, as it's meant for internal use.

Change it to work like `apply` instead of `call`, i.e. to accept an
array instead of varargs, in order to allow passing all arguments to
the super method.

Consistently use the _super function to call super methods.

Introduce a util method `drop` that works like the underscore / lodash
function of the same name to simplify modifying arguments passed to a
super method.

Fix #55

Change-Id: I6fdeda3585bf6933e363995d30ad2b968f6023ac