Let Widget.append method support arrays
The append method accepts widget collections already. However, when
trying to append widgets given as an array (e.g. from a function
parameter) this array would have to be wrapped in a collection.

Change the append method to accept an array of widgets as well.

Cleanup tests for append: keep and initialize children in a shared
beforeEach section.

Change-Id: Ib6a6f3e28a52fc6d2ee5d419460f93c1f5329fd9