Disallow using constructors as factory functions
The ES 2016 specification explicitly demands that constructors throw
a TypeError when called as a function, i.e. without the `new` keyword
(see section 9.2.1 [1]).

If we decide use classes to create widgets, we cannot continue to allow
creating widgets without `new`. Even if we don't, custom widgets that
extend Widget would not allow for it, so we should not do so either.

This reverts commit c2bd17b1.

[1] https://tc39.github.io/ecma262/#sec-ecmascript-function-objects-call-thisargument-argumentslist

Change-Id: Ibbc2ef74715f194048bec438b719e6050091608e