Disable jshint warnings for constructors
The jshint rule "nonew" raises the warning "Do not use 'new'
for side effects" when a constructor is called and the created
instance is not used.

When creating Tabris widgets using constructors, there are
cases that don't require any further references to the created
widget. Even tough these cases are rare, we have some of them
in our snippets and in tests.

Rather than putting jshint exceptions in snippets, it seems
reasonable to remove this check from the jshint config.

Change-Id: I2b19aba1a2b0c676a949ddd928903a836e03bc53