Use new instead of tabris.create in examples
As of Tabris.js 1.7, widgets can be created using constructor
calls instead of `tabris.create`. This syntax is shorter and
easier to remember. While `tabris.create` remains supported
in 1.x, new code should use the new syntax.

Adjust all snippets and examples to use constructors and to
depend on "^1.7.0".

Part of #790

Change-Id: Id990205aba5373c67b87757199742caa0d9728e8