Include polyfill for Promise and fetch
Use the "promise" module from npm, in contrast to other shims
it's self-contained and has a clear structure.

Only require the file "polyfill.js" which does not include the
node specific parts.

Use webpack to create a packed version of the polyfill.

The minified polyfill is about 10k in size, in a built app,
it adds ~~15 ms to the loading time on an iPad 2.

Change-Id: I0856872c0269fb7b5c58e7b3cb13225e7bfd34d1