Generate type definitions in build
As of TypeScript 1.6, the TypeScript compiler can load typings
directly from npm modules.

Add a grunt task that generates typings from the definitions kept
in JSON files and write them directly to the tabris module being
built. Include this task in the grunt "doc" task.

Reference the typings in the package.json of the tabris module.

The files in doc/api/typescript can be used to test the generated
typings. This should later be automated.

Types that don't have a JSON definition are still not generated,
but hard-coded in the generator.

There are still issues with the generated typings, however,
they are already more complete than those currently available
on definitelytyped.org. We can make incremental improvements
from here on.

Fix #814

Change-Id: I8a1844650ce9e4b731875bdcab4d609cbc80c499