Fix TypeScript error when appending Tab to TabFolder
Since TabFolder does not inherit from Composite, calling
`appendTo` with a TabFolder raises an error in TypeScript.

Let TabFolder inherit from Composite. Document that only tabs
are supported as children.

We slightly loose the typing by not enforcing the type of
children in the `append` method. However, it is not enforced
in `appendTo` anyway.

We might consider using generic types to enforce children
types in both, append and appendTo.

#757

Change-Id: I55c99c19cc8e53a5a38e6babff5d44b45427876a