Test generated type definitions in build
On order to discover typescript incompatibilities in our JSON
definitions, run the typescript compiler on the generated type
definitions as recommended in [1].

I've considered using the grunt-typescript task but couldn't
figure out how to configure it to accept the d.ts file. It
always tried to translate the d.ts file which leads to errors.
Moreover, the grunt-typescript module doesn't use the most
recent version of typescript (1.6.x vs. 1.8.x).

Using grunt-exec allows us to specify the exact command,
that can be replicated on the command line.

[1] http://definitelytyped.org/guides/contributing.html

Change-Id: I99d7e7342ffe64a6ca511707d036370e146bd5b0