Insert missing layout data in JavaScript
In case of missing horizontal or vertical layout information,
insert the agreed defaults instead of leaving the resolution
to the native clients.

For the same reason, send SET operations for layoutData even
for widgets with missing layoutData. Do not provide layoutData
for widgets that don't have a parent.

Rename method checkConsistency to normalize, as it amends the
given layoutData.

Adjust affected tests, mostly by making the selection of calls
more specific.

Fix #803

Change-Id: Ia6c1d2083f18a6c7abb12762776e10110920c6d8