Run tests on Node.js instead of Phantom
The grunt-contrib-jasmine module runs tests on Phantom.js in order to
provide a browser environment. Phantom is a huge download, and download
problems have lead to build failures in the past.

Running the tests on Node.js is closer to the target environment of
Tabris (where browser API is partly stubbed). It's also faster than
Phantom and reduces dependencies.

As there is no up-to-date module that integrates plain Jasmine with
Grunt, use the exec task. This allows us to run tests stand-alone.

When we modularize our source code, we can move the imports from the
run-test.js files to the individual tests.

Discard a single test for the module export of tabris. This test would
not work in a Node.js environment. Implicitly test the module export by
assigning it to the global tabris variable.

Change-Id: I846624d0057733edd4e217cc98a06db74e433997