Use lambdas for inner functions in tests
Replace functions in filter(), map(), and expect() with lambda
expressions to reduce noise.

Don't use lambdas for beforeEach, afterEach, describe, etc. as
the test framework might expect these function to have a context
(I've read in the mocha docs that this is discouraged there).

Replace `toThrow(new Error(...))` with `toThrowError(...)`.

Change-Id: I18a5e5cd27fe28ee933922c5e09aaa638aa7eced