Simplify normalized properties
Every property provided to registerType is passed through a normalize
function and added to the _properties object on the widget constructor.

The output of the normalize function does not need to have the same
structure as the input. Simplify the normalized properties objects by
replacing the `access` object with a get and a set function for faster
access.

Do not pass default properties though normalizeProperties before
merging, this only resulted in double normalization.

Do not export the normalizeProperties function anymore.
Test results instead of implementation.

Change-Id: I37beaccdd5a97d9a22caacd5676e68f505e76bdf