Restrict internal util object to function scope
As of 96cbed58, tests are run on node. This requires making the
internal `_` object, that is hidden in tabris.js, available for
tests.

This change accidentally put the `_` in the global scope, which
makes it possible to override it.

Restore the `var _ = ...` to restrict this object to function scope.
Make the `_` object available for tests as `tabris.util`.

Fix #830

Change-Id: Iaef49f0aff5bc709fb5ee46a5d75a7831517009e