Create global window object if not present
The tabris module creates some Web API objects in the global scope.
It does not create a window object, which is expected to be present.

When trying to load the tabris module in Node.js, where window is not
present, an error is thrown. To facilitate testing on Node.js, create
the `window` object if it does not exist.

Change-Id: I0329dc57527cb7b1931edc4389a4f3aef1629b17