TypeScript: Add type definitions and tests for promise and fetch
Add type definitions for Promise based on the existing d.ts file [1]
with some documentation and modifications based on [2] and the API
reference at [3]. Do not add the methods "nodeify" and "denodeify"
mentioned in the API reference.

Add definitions based on the existing whatwg-fetch.d.ts file [4].

The new definitions are copied into the module folder and have to be
explicitly referenced in the tsconfig.json to work.

[1] https://raw.githubusercontent.com/DefinitelyTyped/DefinitelyTyped/14fe4313f4a1cf69add3505a2ab1dbc690dc2116/promise/promise.d.ts
[2] https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/es6-promise/es6-promise.d.ts
[3] https://www.promisejs.org/api/
[4] https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/whatwg-fetch/whatwg-fetch.d.ts

Change-Id: I662442ff8130815f43cd880f02a85be7ccba8fc0