Add support for crypto.getRandomValues()
For cryptographic applications, provide a source for cryptographically
secure random numbers.

Implement the RandomSource API [1] and make it available as global
object `crypto` for compatibility with browser's `window.crypto`.

Since the iOS client cannot modify parameters, pass the number of
requested bytes to the client. Clients are expected to return an
array of the requested length, filled with random integer numbers
in the range [0 .. 255].

Fix #834

[1] https://developer.mozilla.org/en-US/docs/Web/API/RandomSource

Change-Id: I8e84e8d4000a8bc85e50114600d493a28e0f3b2e