Skip dispose checks for event methods
Adding or removing listeners on a disposed widget doesn't do any
harm, native events won't just be fired anymore.

In order to reduce the risk of occasional errors in asynchronous
application code, do not throw exceptions when calling any of these
methods on a disposed widget.

This also reduces the need for disposed-checks in asynchronous
application code.

The disposed errors haven't been documented, no need for changes to
the documentation.

Change-Id: Ic49cf6c3387e5426b0adaec443fbc32cb6a89918