Revise Web Storage implementation
Prevent window.Storage from being instantiated.

Rename constructor tabris.WebStorage to tabris.Storage to match
the Web API [1]. This constructor is not documented, so it's not
a breaking change.

In localStorage, keep the proxy in an instance field.
Use defineProperty() to make this._proxy non-enumerable, i.e. prevent
it from showing up in Object.keys().

Add window.Storage and window.localStorage on load event and unit test
the actual instance.

[1] https://developer.mozilla.org/en/docs/Web/API/Storage

Change-Id: Iea5e571cdf8a9bcddcd29cf56a0d47e7715e2aaa