Replace Shell and Display with a Composite, make UI a plain object
The composite is available as `tabris.ui.content` and will serve as the
root element for all other widgets except Drawer. It is marked with a
property "root" set to `true` in its create operation.

Replace Tabris.UI with a simple plain object that has only the
read-only "content" field. Remove all references to the old UI
implementation from Page, Action, SearchAction and Drawer.
Do not touch PageSelector since it's too tightly bound to the
old API and has to be adjusted later. 

Adjust one snippet (layout) to demonstrate the new API. No documentation
yet since all changes may not be final.

Fix #946

Change-Id: Ia8bfb6746f5e7cbb85c7e845e5e771d895890732