Enable real unit tests with mocha
Currently, we run all tests against the final tabris build. To test
internals, we have to expose them on the tabris module.
Now that the source code is organized in modules, we can enable unit
tests on a per-module basis, without having to expose internals.

Accessing the exports of a module requires transpiling the ES6 import
and export statements during a test run. Use mocha, which is able to
transpile source and test code transparently.

Keep jasmine for existing *.spec.js files and use mocha for new
ones. To distinguish between jasmine and mocha tests, use the
suffix *.test.js instead of *.spec.js. This allows to migrate
existing tests to mocha one-by-one.

Adjust the test for ProxyStore for mocha as a proof of concept.

Change-Id: I0a6a179a052e7443629f1978222e761415555296