Remove support for browser tests
Test can not be run in the browser anymore, since the browser can't
handle ES6 import/export statements. With debugging support available
in VS code, this is not so much of an issue anymore.

Remove HTML files used for browser tests, as well as FakeTabrisModule
which was only needed in the browser.

Change-Id: I71bea679d67d615a5d7cbf090c28b938f116dcc8