Prevent conflict in the "image" Page property
Rename property to "icon" to prevent conflict with the "image" property
used in the PageSelector.

We can safely remove the "image" property and don't have to introduce
a "icon" to exist side-by-side with "image", as "image" was only
consumed in Tabris.js by the PageSelector API and is as of 1.9 not
supported by the clients.

Change-Id: I57855951e1e6a9bdec2ed49d1cd3ef89eaff3c7b