Prevent page image property name conflict
Rename property to "icon" to prevent conflict with the "image" property
used in the PageSelector.

We can safely remove the property and don't have to introduce a new one
to exist side-by-side with "image", as "image" was only consumed in
Tabris.js by the PageSelector API and is as of 1.9 not supported by the
clients.