Turn DOMEvent and DOMDocument into modules
Rename DOMEvents to DOMEvent and export the constructor. Also export
methods to apply document and event support to a target.

Move the code to apply document on window to main.js.

Remove checks for existing window and module globals when applying
those APIs in main.js, as we don't test main.js and globals are
available at runtime.

Change-Id: I622146f037b69ae1c9e4358c3fe06412e2d04874