Move type name into Proxy/Widget config
The name of a Proxy or Widget subclass is needed for type selectors.

Since the extend() methods on Proxy and Widget don't register the
constructors on the tabris object anymore, the type name is just
another configuration property.

Remove the first parameter in Proxy.extend() and Widget.extend()
and move the type into the config. Call it _name, because _type
is already used for the native type.

Change-Id: I3bf7637802c614d122de85b4f99e93e12a2169ab